pyproject.toml
[tool.poetry]
name = "tensorexperiments"
version = "0.1.0"
description = "Tensorflow environment setup with Pyenv and Poetry for Macbook Air M1 with ARM-64 architecture"
authors = ["Alex Watson <alexfromapex@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
tensorflow-macos = "2.8.0"
absl-py = "1.0.0"
astunparse = "1.6.3"
CacheControl = "0.12.11"
cachetools = "5.0.0"
cachy = "0.3.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
cleo = "0.8.1"
clikit = "0.6.2"
crashtest = "0.3.1"
Cython = "0.29.28"
distlib = "0.3.4"
filelock = "3.6.0"
flatbuffers = "2.0"
gast = "0.5.3"
google-auth = "2.6.6"
google-auth-oauthlib = "0.4.6"
google-pasta = "0.2.0"
grpcio = "1.46.1"
h5py = "3.6.0"
html5lib = "1.1"
idna = "3.3"
importlib-metadata = "4.11.3"
keras = "2.8.0"
Keras-Preprocessing = "1.1.2"
keyring = "23.5.0"
libclang = "14.0.1"
lockfile = "0.12.2"
Markdown = "3.3.7"
msgpack = "1.0.3"
numpy = "1.22.3"
oauthlib = "3.2.0"
opt-einsum = "3.3.0"
packaging = "20.9"
pastel = "0.2.1"
pexpect = "4.8.0"
pkginfo = "1.8.2"
platformdirs = "2.5.2"
poetry = "1.1.13"
poetry-core = "1.0.8"
protobuf = "3.20.1"
ptyprocess = "0.7.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pylev = "1.4.0"
pyparsing = "3.0.9"
requests = "2.27.1"
requests-oauthlib = "1.3.1"
requests-toolbelt = "0.9.1"
rsa = "4.8"
shellingham = "1.4.0"
six = ">=1.15.0,<1.16.0"
tensorboard = "2.8.0"
tensorboard-data-server = "0.6.1"
tensorboard-plugin-wit = "1.8.1"
termcolor = "1.1.0"
tf-estimator-nightly = "2.8.0.dev2021122109"
tomlkit = "0.10.2"
typing-extensions = "4.2.0"
urllib3 = "1.26.9"
virtualenv = "20.14.1"
webencodings = "0.5.1"
Werkzeug = "2.1.2"
wrapt = "1.14.1"
zipp = "3.8.0"
tensorflow-metal = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tf = 'tensorexperiments.tf:main'